carve off datastore factory to its own file, refactor reference fetching
