+++ tests/seechange_config_test.yaml
@@ -22,5 +22,4 @@ archive:
 
 subtraction:
   method: zogy
-  refset: test_refset_ptf
